+++ org.csstudio.javafx.rtplot.test/src/org/csstudio/javafx/rtplot/ImagePlotDemo.java
@@ -14,6 +14,7 @@ import java.util.logging.Handler;
 import java.util.logging.Level;
 import java.util.logging.Logger;
 
+import org.csstudio.javafx.rtplot.internal.ImagePlot;
 import org.diirt.util.array.ArrayDouble;
 import org.diirt.util.array.ListDouble;
 
@@ -67,9 +68,9 @@ public class ImagePlotDemo extends Application
         for (Handler handler : Logger.getLogger("").getHandlers())
             handler.setLevel(level);
 
-        final RTImagePlot plot = new RTImagePlot();
+        final ImagePlot plot = new ImagePlot();
 
-        plot.setColorMapping(RTImagePlot.RAINBOW);
+        plot.setColorMapping(ImagePlot.RAINBOW);
 
         plot.setAutoscale(true);
         plot.setValueRange(0.0, 2.0);
@@ -86,7 +87,6 @@ public class ImagePlotDemo extends Application
             show_colorbar = ! show_colorbar;
             plot.showColorMap(show_colorbar);
         }, 5000, 5000, TimeUnit.MILLISECONDS);
-
 
 		final Pane root = new Pane(plot);
 		plot.widthProperty().bind(root.widthProperty());
